BRTM development
A propagation of virtual packets is now processed through the CCD chip.
A very basic testing has been processed. It may be working.

+++ save_rates.f90
@@ -8,7 +8,7 @@ INTEGER                                 :: cur_line
 INTEGER                                 :: indexe, indexi
 INTEGER                                 :: cur_mgi
 INTEGER                                 :: lower_level, upper_level
-INTEGER                                 :: stat_weight_l, stat_weight_u
+DOUBLE PRECISION                        :: stat_weight_l, stat_weight_u
 DOUBLE PRECISION                        :: ROverV, corrFactor, cur_kappa, cur_radius, cur_rho
 DOUBLE PRECISION                        :: constanta
 DOUBLE PRECISION                        :: cur_wavelength
